Deduplicate mutator controller logic
Fixes #1449
