+++ pkg/controller/mutators/core/controller_suite_test.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package assignmetadata
+package core
 
 import (
 	"context"
